refactor: create routes directory
